+++ changelog.md
@@ -20,7 +20,7 @@ Patch: Everything else → bumps patch version (1.0.0 → 1.0.1)
 
 ### Fixed
 
--
+- fixed issue with markdown viewer not working on news and events
 
 ## [2.5.0] - 2025-07-06